moved PeaqCurrencyAdapter to common-runtime, removed DealWithFees
