+++ runtime/common/Cargo.toml
@@ -17,6 +17,7 @@ frame-system = { workspace = true, default-features = false }
 orml-tokens = { workspace = true, default-features = false }
 orml-traits = { workspace = true, default-features = false }
 pallet-transaction-payment = { workspace = true, default-features = false }
+pallet-block-reward = { path = "../../pallets/block-reward", default-features = false }
 peaq-primitives-xcm = { path = "../../primitives/xcm", default-features = false }
 sp-core = { workspace = true, default-features = false }
 sp-runtime = { workspace = true, default-features = false }
@@ -36,6 +37,7 @@ std = [
 	"orml-tokens/std",
 	"orml-traits/std",
 	"pallet-transaction-payment/std",
+	"pallet-block-reward/std",
 	"peaq-primitives-xcm/std",
 	"sp-core/std",
 	"sp-runtime/std",
